requirements.txt
black==23.9.1
faiss-cpu==1.7.4
langchain==0.0.295
numpy==1.26.0
openai==0.28.0
pydantic==2.3.0
pypdf==3.16.1
python-dotenv==1.0.0
regex==2023.8.8
tqdm==4.66.1
